various change

+++ src/bb_methods_classes_intro/bb_StaticMethodTest/MyMath.java
@@ -1,7 +1,7 @@
 package bb_methods_classes_intro.bb_StaticMethodTest;
 
 public class MyMath {
-	static double myMax ( float aa , float bb ) {
+	static double myMax ( double aa , double bb ) {
 		System.out.println( "using double,double myMax ") ;
 		if ( aa > bb ) {
 			return aa ;
